# Documents/SUNNU_MBAYE_MI/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/commander": {
            "post": {
                "summary": "Passer une commande",
                "description": "Passer une commande.",
                "operationId": "30391e8606a0ccc5194fc1cec56bd789",
                "responses": {
                    "200": {
                        "description": "Commande passée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commande passée avec succès"
                                        },
                                        "commende_id": {
                                            "type": "integer"
                                        },
                                        "total_prix": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Le panier est vide"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/afficheCommende": {
            "get": {
                "summary": "Afficher les commandes de l'utilisateur",
                "description": "Afficher les commandes de l'utilisateur.",
                "operationId": "e063cb6455982a11f3379e8db24d3e1e",
                "responses": {
                    "200": {
                        "description": "Liste des commandes de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "commende": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Commende"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/voirPlus/{commende_id}": {
            "get": {
                "summary": "Afficher les détails d'une commande",
                "description": "Afficher les détails d'une commande.",
                "operationId": "3d6daf762b9e98a58a2b568826870e56",
                "parameters": [
                    {
                        "name": "commende_id",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la commande",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "commende": {
                                            "$ref": "#/components/schemas/Commende"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Veillez vous connecter d'abord"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supprimerCommende/{commende_id}": {
            "delete": {
                "summary": "Supprimer une commande",
                "description": "Supprimer une commande.",
                "operationId": "6b1017c09c0a6ff2f475085b6a666a35",
                "parameters": [
                    {
                        "name": "commende_id",
                        "in": "path",
                        "description": "ID de la commande à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Commande supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Veillez vous connecter d'abord"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/annulerLivraison/{$commende_id}": {
            "put": {
                "summary": "Annuler la livraison d'une commande",
                "description": "Annuler la livraison d'une commande.",
                "operationId": "06e5957ac25e96d65ad816428f0d0a3a",
                "parameters": [
                    {
                        "name": "commende_id",
                        "in": "path",
                        "description": "ID de la commande à annuler la livraison",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livraison annulée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Livraison annulée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/livraisonTerminer/{commende_id}": {
            "put": {
                "summary": "Marquer la livraison d'une commande comme terminée",
                "description": "Marquer la livraison d'une commande comme terminée.",
                "operationId": "c98a82d12232857aff2164e5c442abce",
                "parameters": [
                    {
                        "name": "commende_id",
                        "in": "path",
                        "description": "ID de la commande à marquer comme terminée",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livraison terminée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Livraison terminée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}